Module 8: Lab Activity – Booleans
A Boolean Algebra is the basis of all modern computer arithmetic, where there are True and False Boolean values. It is important in relating to programming because if any block’s Boolean expression becomes true, that block will run, and the block will not run in case of false. 

Game Example from How to Think Like a Computer Scientist:
For example, suppose we can slay the dragon only if our magic lightsabre sword is charged to 90% or higher, and we have 100 or more energy units in our protective shield. We find this fragment of Python code in the game:
	if not ((sword_charge >= 0.90) and (shield_energy >= 100)):
     print("Your attack has no effect, the dragon fries you to a crisp!")
else:
     print("The dragon crumples in a heap. You rescue the gorgeous princess!") 
de Morgan’s laws, together with the logical opposites, would let us rework the condition in a (perhaps) easier to understand way like this:
	if (sword_charge < 0.90) or (shield_energy < 100):
     print("Your attack has no effect, the dragon fries you to a crisp!")
else:
     print("The dragon crumples in a heap. You rescue the gorgeous princess!") 
We could also get rid of the not by swapping around the then and else parts of the conditional. So here is a third version, also equivalent:
	if (sword_charge >= 0.90) and (shield_energy >= 100):
     print("The dragon crumples in a heap. You rescue the gorgeous princess!") else:
     print("Your attack has no effect, the dragon fries you to a crisp!")